main/requirements.txt
streamlit
sqlite3
Pillow
google-generativeai
wikipedia-api
re
gtts
googletrans
SpeechRecognition